Fix image keys when image URLs contain query params
If an image URL contained a query param, those params would end up
in the key used for the Phaser image cache.

This resulted in cache misses, because the plugin doesn't expect
these query params.

This commit fixes the issue, by stripping any query params when composing
the image key.